[refactor] RadiusOrangeButton.jsx 버튼 스타일 변경 및 사용되지 않는 파일 삭제
