--- src/styles/BoardButtonStyle.js
@@ -1,3 +1,0 @@
-import styled from "styled-components";
-
-export const BoardButtonContainer = styled.div``;
