+++ src/styles/BoardStyle.js
@@ -86,13 +86,4 @@ export const BoardContainer = styled.div`
     line-height: 1.5;
     box-shadow: rgba(50, 50, 93, 0.25) 0px 2px 5px -1px, rgba(0, 0, 0, 0.3) 0px 1px 3px -1px;
   }
-
-  button {
-    background-color: #fc913a;
-    color: white;
-    border: none;
-    padding: 10px;
-    margin: 0 20px;
-    cursor: pointer;
-  }
 `;
